chore: add clippy checking to CI
This patch adds a `clippy` check to CI, and fixes related issues
in the source code.

There were a few small complexities to making this change. The first
is that some of the tests would do `assert!(false)`, and `clippy`
would (rightly) point out that one should just panic there.

The second is that `pyo3` was generating some code in a macro that
clippy didn't like, but since it was in the macro, it couldn't be
ignored either. This led to the need to upgrade `pyo3`, which may
actually have impactful changes. The tests still run, but I have
no real visibility into the downstream consumers. The `abi3-py36`
feature is gone, as 3.6 has been EOL since 2021, so that may have
some lasting consequences.